+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - Added advanced option to block messages via notification
 - Added option to change the way response messages are sent
 - Switched response message delays from individual to cumulative
+- Added CommandKeys integration for response messages